refactoring api-action
