+++ project/src/components/review/review.tsx
@@ -1,5 +1,5 @@
 import { ChangeEvent, FormEvent, Fragment, useState } from 'react';
-import { RatingRewiev } from '../../const';
+import { RatingRewievs } from '../../const';
 import { useAppDispatch } from '../../hooks';
 import { sendReviewAction } from '../../store/api-action';
 
@@ -41,7 +41,7 @@ const ReviewForm = ({id}: Props) => {
     <form className="reviews__form form" action="#" method="post" onSubmit={handleSubmit}>
       <label className="reviews__label form__label" htmlFor="review">Your review</label>
       <div className="reviews__rating-form form__rating">
-        {RatingRewiev.map((mark) => (
+        {RatingRewievs.map((mark) => (
           <Fragment key={mark.id}>
             <input
               className = "form__rating-input visually-hidden"
