+++ project/src/pages/main-page/main-page.tsx
@@ -7,6 +7,7 @@ import Sorting from '../../components/sorting/sorting';
 import { getCity } from '../../store/data/data.selector';
 import { getSelectCardId } from '../../store/data/data.selector';
 import { getCityOffers } from '../../store/data/data.selector';
+import EmptyCtiy from '../../components/empty-ctiy/empty-city';
 
 function getTextByCount(count: number, city: string): string {
   const pluralRules = plural.select(count);
@@ -18,7 +19,7 @@ function getTextByCount(count: number, city: string): string {
   }
 }
 
-const MainPages = () =>{
+const MainPages = () => {
 
   const city = useAppSelector(getCity);
   const points = useAppSelector(getCityOffers);
@@ -33,19 +34,21 @@ const MainPages = () =>{
           <Cities />
         </section>
       </div>
-      <div className="cities">
-        <div className="cities__places-container container">
-          <section className="cities__places places">
-            <h2 className="visually-hidden">Places</h2>
-            <b className="places__found">{getTextByCount(points.length, city.name)}</b>
-            <Sorting />
-            <ProductList offers={points} className='cities__places-list places__list tabs__content' />
-          </section>
-          <div className="cities__right-section">
-            <Map city={city} offers={points} pointId={pointId} className='cities__map map' />
+      {points.length !== 0 ?
+        <div className="cities">
+          <div className="cities__places-container container">
+            <section className="cities__places places">
+              <h2 className="visually-hidden">Places</h2>
+              <b className="places__found">{getTextByCount(points.length, city.name)}</b>
+              <Sorting />
+              <ProductList offers={points} className='cities__places-list places__list tabs__content' />
+            </section>
+            <div className="cities__right-section">
+              <Map city={city} offers={points} pointId={pointId} className='cities__map map' />
+            </div>
           </div>
         </div>
-      </div>
+        : <EmptyCtiy />}
     </main>
   );
 };
