+++ project/src/types/offers.ts
@@ -72,19 +72,6 @@ export type ReviewOfferCard = {
   date: string;
 };
 
-export type CommentData = {
-  id: number;
-  rating: number;
-  review: string;
-}
-
-export type CommentFormData = {
-  rating: number;
-  comment: string;
-}
-
-export type Comments = Comment[];
-
 export type ReviewOfferCards = ReviewOfferCard[];
 
-export type OfferGroup = (OfferCard | OfferCards | ReviewOfferCards) [];
+export type OffersGroup = (OfferCard | OfferCards | ReviewOfferCards) [];
